90_openvino_yolo_experiment/yolo_consumer/requirements.txt
confluent-kafka
Pillow
torch
opencv-python
ultralytics
bezier
openvino

gitpython
gitdb
smmap
setuptools==68.2.2

onnxruntime
onnx>=1.12.0
onnxslim==0.1.34


# requirements: Ultralytics requirements ['onnx>=1.12.0', 'onnxslim==0.1.34', 'onnxruntime'] not found, attempting AutoUpdate...
